On Unix platforms zip Boost headers used by system tests
Some of the system tests need Boost.MPL to be available on the include path.
On most newly installed systems the Boost headers are not available. Copying
the headers required by the system tests to the target host and adding them to
the include path when the tests are executed is required to make the tests pass.
Zipping the required headers together makes this easier.