Do all documentation generation from reference_gen
reference_gen has been a special purpose tool for generating documentation for
Metashell. Keeping it and replace_part separate means that these tools need to
be combined using the build system (CMake). Doing that (and maintaining it) in a
multi-platform way is more complicated than implementing the logic in C++ and
launching it from the build system.